requirements.txt
fastapi[standard]
openai>=1.35.0
python-dotenv>=1.0.1
sqlalchemy>=2.0.30
psycopg2>=2.9.9
alembic>=1.13.2
aiofiles>=24.1.0
asyncio>=3.4.3
boto3 >=1.34.143
python-jose[cryptography] 
passlib[bcrypt]
bcrypt
fastapi-socketio >=0.0.10
google-cloud-aiplatform >=1.60.0
